+++ packages/codemod/src/utils/additionalContent.ts
@@ -2,7 +2,7 @@ import * as fs from 'fs-extra';
 
 export default function (filePath, addContent, templatePath) {
   if (fs.existsSync(filePath)) {
-    let fileContent = fs.readFileSync(filePath);
+    let fileContent = fs.readFileSync(filePath, 'utf8');
     if (!fileContent.includes(addContent)) {
       fileContent += `\n${addContent}\n`;
       fs.writeFileSync(filePath, fileContent);