+++ packages/plugin-rax-web/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v1.4.6
+
+- Fix: document live load
+- Feat: support sdk combo
+
+## v1.4.5
+
+- Fix: throw document build error
+
+## v1.4.4
+
+- Fix: `doctype: null`
+
 ## v1.4.3
 
 - Chore: optimize hot reload experience
@@ -26,7 +39,7 @@
 
 ## v1.3.8
 
-- Fix: rax render won't override original content with initialHTML is undefined 
+- Fix: rax render won't override original content with initialHTML is undefined
 
 ## v1.3.7
 
@@ -65,4 +78,3 @@
 
 - Refactor: document
 - Feat: support config `hash` in build.json
-