+++ packages/rax-babel-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.0.5
+
+- Feat: add config option `forceEnableCSS` support
+
 ## v1.0.4
 
 - Fix: JSX Plus docs url in console